fix(swarm): reuse allocated ports after restart (#1056)
Description
---
uses previously allocated ports when restarting an instance

Motivation and Context
---
If an instance was stopped and started, it would receive new ports,
causing other instances that may depend on the previous ports to break.
This PR reuses the previously assigned ports for the instance on
restart.

How Has This Been Tested?
---
Manually 

What process can a PR reviewer use to test or verify this change?
---
Click stop then start on an instance, ports should remain the same.

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify