feat(consensus): optionally configure max block time (#1097)
Description
---
Update consensus constants to include the (maximum) base block time, a
component of the GST before a leader fails.

Resolves #1088 

Motivation and Context
---

Currently it is hard coded to 10 seconds. This will make it
configurable.

How Has This Been Tested?
---

Unit and integration tests.

What process can a PR reviewer use to test or verify this change?
---


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify