feat: new instruction type AssertBucketContains (#1106)
Description
---
* New instruction `AssertBucketContains` and associated utils and errors
* New workspace action for the engine to handle the new instruction
* Updated protobuf definitions with the new instruction type
* Updated TypeScript bindings with the new instruction type
* Bumped TypeScript bindings package version to trigger a npm publish
* Made `tari_dan_wallet_web_ui` npm dependencies with bindings and
clients use the npm registry instead of local dependencies, to avoid
chicken-and-egg situations.

Motivation and Context
---
We want a new type of transaction instruction to execute asserts.
Asserts can be useful in a variety of ways, making transactions more
secure and easier to review by the user. For example, it could allow to
implement a maximum slippage in a decentralized exchange.

The main use case is to ensure that the return bucket of a component
call invocation contains the expected resource and amount of tokens,
hence the new instruction is called `AssertBucketContains`. The fields
of the new instruction type are:
* `key`: workspace reference of the bucket that we want to assert
* `resource_address`: the resource address that the bucket must contain
* `min_amount`: the minimum amount of tokens that the bucket must
contain

How Has This Been Tested?
---
New engine unit tests that cover all scenarios

What process can a PR reviewer use to test or verify this change?
---
Send a transaction with the new instruction type

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify