fix(engine): finalize errors (e.g. dangling bucket) reject the entire transaction (#1109)
Description
---
Extracted the `validate_finalized` check invocation from the
`StateTracker` into the `TransactionProcessor`, to be done automatically
at the end of all `instructions` and `fee_instructions`.

Motivation and Context
---
Currently, when the main instruction set has a dangling bucket (or
another finalizing error), the whole transaction is rejected. What we
want instead is only the main transaction being rejected, and the fees
are still paid (`AcceptFeeRejectRest`)

How Has This Been Tested?
---
New unit test (`dangling_bucket_pay_fees`) in the engine to reproduce
this scenario

What process can a PR reviewer use to test or verify this change?
---
Reproduce a dangling bucket error and check that the fees are paid

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify