feat(watcher): add mattermost notifier for channel alerting (#1126)
Description
---
Adds alerting for a channel on Mattermost with a generalized notifier
implementation. Next step is to add Telegram support.

How Has This Been Tested?
---
Create a Mattermost channel and log in to the server using the api to
obtain credentials. After updating the config file (with both
`channel_id` and `credentials`), run `tari_swarm_daemon` together with
the `tari_watcher`, and go through the initial registration step,
warning on close to expiry, and re-registration. Observe messages posted
in the channel.