feat(consensus)!: add sidechain id in the genesis block (#1135)
Description
---
* Created the definition of a `ExtraData` (following Bitcoin and
Ethereum naming) type for storing additional information inside blocks:
    * It contains a key-value map of arbitrary binary data
* For now only a key exists for sidechain ids, but it can be extended in
the future
* The binary value max size is limited by definition via `MaxSizeBytes`
(copied from the minotari codebase)
* Updated the definition for `Block`, `ParkedBlock` and
`ForeignProposal` to include an optional `extra_data` field
* Updated the SQLite storage for the new field
* Updated the protobuf definitions with the new field
* Updated the typescript bindings
* Genesis blocks now encode the sidechain id ( if present in the config)
in the new `extra_data` field
* New `check_sidechain_id` validation in consensus block validation, to
check the sidechain id in the extra data of genesis blocks

Motivation and Context
---
We should have the sidechain id in the genesis block for the shard and
epoch. Including it in all blocks would be wasteful, so only genesis
blocks should include it.

To make it more general for the future, we want to create an optional
field in blocks to store any type of data, and use it to store the
sidechain id.

How Has This Been Tested?
---
Manually with `tari_swarm`:
* Setting up a `sidechain_deployment_key` in the `minotari_wallet`
* Settiup up a `validator_node_sidechain_id` in `validator_node`
* Setting up `sidechain_id` in the `indexer`
* Performing transactions to advance the network and inspecting the
blocks in the database

Also tested with no sidechain

What process can a PR reviewer use to test or verify this change?
---
See previous section

Breaking Changes
---

- [ ] None
- [ ] Requires data directory to be deleted
- [x] Other - Network reset as the block definition and validation has
changed